# docs/swagger.yaml
basePath: /api/v1/
definitions:
  domain.Empolyee:
    properties:
      created_at:
        type: string
      id:
        type: integer
      name:
        type: string
      position:
        type: string
      salary:
        type: number
      updated_at:
        type: string
    type: object
  domain.EmpolyeeRequestDto:
    properties:
      id:
        type: integer
      name:
        type: string
      position:
        type: string
      salary:
        type: number
    type: object
externalDocs:
  description: OpenAPI
  url: https://swagger.io/resources/open-api/
host: localhost:8000
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: This is a sample server celler server.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Empolyee Example API
  version: "1.0"
paths:
  /empolyees:
    post:
      consumes:
      - application/json
      description: Create an new empolyee record.
      parameters:
      - description: Authorization
        in: header
        name: Authorization
        required: true
        type: string
      - description: User data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/domain.EmpolyeeRequestDto'
      produces:
      - application/json
      responses:
        "201":
          description: Empolyee created successfully
          schema:
            type: string
        "400":
          description: Invalid request
          schema:
            type: string
      summary: Create new Empolyee
      tags:
      - Empolyees
  /empolyees/{id}:
    get:
      consumes:
      - application/json
      description: Retrieve employee details by employee ID
      parameters:
      - description: Authorization
        in: header
        name: Authorization
        required: true
        type: string
      - description: Employee ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.Empolyee'
        "400":
          description: Bad Request
          schema: {}
      summary: Get an employee by ID
      tags:
      - employees
  /ping:
    get:
      consumes:
      - application/json
      description: do ping
      parameters:
      - description: Authorization
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
      summary: ping example
      tags:
      - Ping
securityDefinitions:
  BasicAuth:
    type: basic
swagger: "2.0"